:sparkles: add backend that will serve API
Boy, was this a hassle. Writing the actual Express server was fine,
but getting it setup properly with `now` and adding Google's private key
as a secret variable took forever. THEN when that was solved I couldn't
for the life of me get the server deployed using `now` v2.

Ultimately ended up deciding to deploy the backend using v1 and will
have to read up on `now` lambdas and think about rewriting this as
serverless functions.